finanace calculator

+++ python_introduction/finance_calculator.py
@@ -1,5 +1,5 @@
 monthly_income = int(input("Enter your monthly income: "))
-monthly_expenses = int(input("Enter your total monthly expenses: "))
+monthly_expenses = float(input("Enter your total monthly expenses: "))
 
 projected_saving = monthly_income * 12 + (monthly_income * 12 * 0.05)
 print(f"Your monthly savings are ${monthly_income}.")
